actions/defaultChecker 없이 reducers/defaultChecker 만 써서 구현.
