+++ app/components/WifiStatus.js
@@ -2,7 +2,8 @@
 import React, { Component } from 'react';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import * as WifiCheckerActions from '../actions/defaultChecker';
+// * ===> actionCreators 로 바꿈. actionCreators function들만 묶은 오브젝트.
+import { actionCreators as WifiCheckerActions } from '../reducers/defaultChecker';
 import { firebase } from '../utils/firebase';
 import { Link } from 'react-router-dom';
