+++ app/containers/LoginPage.js
@@ -1,7 +1,8 @@
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import Login from '../components/Login';
-import * as LoginActions from '../actions/defaultChecker';
+// * ===> actionCreators 로 바꿈. actionCreators function들만 묶은 오브젝트.
+import { actionCreators as LoginActions } from '../reducers/defaultChecker';
 
 function mapStateToProps(state) {
     console.log('[LoginPage.js]','mapStateToProps', state);
